printk: Add generic printk operation
Printk used to be platform-specified implementation. But
in the view of reusability, that's bad for maintainability.
This patch extracts common flow of printk and use UART driver
to initialize default UART to work for printk.
