+++ platform/f429disco/init.c
@@ -26,18 +26,3 @@ __weak void __platform_halt(void)
     for (;;)
         ;
 }
-
-void __printk_init(void)
-{
-    usart_init();
-}
-
-void __printk_putchar(char c)
-{
-    if (c == '\n')
-        __printk_putchar('\r');
-
-    while (!(USART1->SR & USART_SR_TXE))
-        ;
-    USART1->DR = (0xff) & c;
-}
